// src/app/model/EessAndProducto.ts
export interface EessAndProducto {
    ideess: string;
    ruteess: string;
    codigogilbarco: string;
    nombrefantasia: string;
    razonsocial: string;
    comuna: string;
    direccion: string;
    productno: 0;
    productname: string;
    modeno: 0;
    modename: string;
    productum: string;
    umshort: string;
    price: 0;
    iddsps: 0;
    dspdeviceid: 0;
    idfps: 0;
    fpdeviceid: 0;
    pumpno: 0;
    nozzleno: 0;
    logicalid: 0;
    // 2/4/2021
    frontend_productname: string;
}